split values src/dst into values.lua #1
